requirements.txt
dm-control==1.0.15
gym==0.26.2
h5py==3.10.0
imageio==2.31.5
matplotlib==3.7.3
mpi4py==3.1.5
numpy==1.24.3
opencv-python==4.8.1.78
pandas==2.0.3
Pillow==10.1.0
pyparsing==3.1.1
pytorch-lightning==1.6.4
requests==2.31.0
scipy==1.10.1
Shimmy==1.3.0
six==1.16.0
stable-baselines3==2.1.0
tensorboard==2.13.0
tensorflow==2.13.1
torch==2.1.0
torchmetrics==1.2.0
tqdm==4.66.1
urllib3==2.0.7
wandb==0.15.12